Move control appearance components into a subdirectory
